Feature/globebrowsing (#334)
Layer support for globe browsing:

Add layers using the function openspace.globebrowsing.addLayer
Delete layers using openspace.globebrowsing.deleteLayer
Layer type does not necessarily have to be of tile type. For example solidcolor does not use tiles
Blend modes for layers are Normal, Add, Subtract, Multiply, Color
Layer adjustments to affect layers. The current only active one is chroma key to cut out a color from the layer. Transfer functions or clipping masks are examples of layer adjustments for the future.
Support for adding layer specifications for quickly accessing GIBS layers:
openspace.globebrowsing.createGibsGdalXml
openspace.globebrowsing.createTemporalGibsGdalXml
The arguments for these functions are currently strings. Would it be better to use a lua dictionary?
No data values for height layers are correctly regarded (can be seen on Earth. No longer bumps on the poles)
Other minor things:

Worked a bit on point globe to render globes at large distances. Currently not in use and doesn't have anything to do with the other things.
Concurrent job manager takes a thread pool as argument and not a pointer to one. This is because the concurrent job manager needs to have ownership of the thread pool for correct deinitialization. Will cause breaking change for users of concurrent job manager if merged in to master.


* Add ability to add layers programatically.

* Clean up

* Fix order of deletion in concurrent job manager and clean up

* Can create by level tile provider with empty dictionary.

* Add script to add GIBS datasets.

* Start working with layer adjustment

* Update mod files

* More work on point globe

* Add script to create temporal GIBS datasets.

* Update temporal tile provider to be able to take gdal descriptions without file path.

* Add adjustment property to layers.

* Rename adjustment layer

* Add adjustment code to all layer groups

* Remove caching of gdal datasets due to cluttering of folders

* Document layer support

* Update Mars mod

* Make Mercury great again.

* Cleanup and add blend mode Color

* Enable setting of layeradjustment and blend mode from mod files.

* No more use for grayscale color overlays. Use grayscale layer with color blend mode instead.

* Clean up mod files

* Clean up

* Clean up

* No need for grayscale layers. Reading grayscale in to rgb instead for color layers.

* Remove unused layer groups

* Correctly read to grayscale layers

* Update globe mod files

* Rename ColorOverlays to Overlays.

* Clean up

* Clean up

* Solve compilation error